+++ watertap/costing/watertap_costing_package.py
@@ -12,8 +12,6 @@
 
 import pyomo.environ as pyo
 
-from pyomo.util.calc_var_value import calculate_variable_from_constraint
-
 from idaes.core import declare_process_block_class
 from idaes.core.base.costing_base import register_idaes_currency_units
 
@@ -55,58 +53,5 @@ class WaterTAPCostingData(WaterTAPCostingBlockData):
         # fix the parameters
         self.fix_all_vars()
 
-    def build_process_costs(self):
-        # add total_captial_cost and total_operating_cost
-        self._build_common_process_costs()
-
-        self.total_capital_cost_constraint = pyo.Constraint(
-            expr=self.total_capital_cost
-            == self.factor_total_investment * self.aggregate_capital_cost
-        )
-
-        self.maintenance_labor_chemical_operating_cost = pyo.Expression(
-            expr=self.factor_maintenance_labor_chemical * self.aggregate_capital_cost,
-            doc="Maintenance-labor-chemical operating cost",
-        )
-
-        self.total_fixed_operating_cost = pyo.Expression(
-            expr=self.aggregate_fixed_operating_cost
-            + self.maintenance_labor_chemical_operating_cost,
-            doc="Total fixed operating costs",
-        )
-
-        self.total_variable_operating_cost = pyo.Expression(
-            expr=(
-                self.aggregate_variable_operating_cost
-                + sum(self.aggregate_flow_costs[f] for f in self.used_flows)
-                * self.utilization_factor
-            )
-            if self.used_flows
-            else self.aggregate_variable_operating_cost,
-            doc="Total variable operating cost of process per operating period",
-        )
-
-        self.total_operating_cost_constraint = pyo.Constraint(
-            expr=self.total_operating_cost
-            == (self.total_fixed_operating_cost + self.total_variable_operating_cost),
-            doc="Total operating cost of process per operating period",
-        )
-
-        self.total_annualized_cost = pyo.Expression(
-            expr=(
-                self.total_capital_cost * self.capital_recovery_factor
-                + self.total_operating_cost
-            ),
-            doc="Total annualized cost of operation",
-        )
-
-    def initialize_build(self):
-        calculate_variable_from_constraint(
-            self.total_capital_cost, self.total_capital_cost_constraint
-        )
-        calculate_variable_from_constraint(
-            self.total_operating_cost, self.total_operating_cost_constraint
-        )
-
-        for var, con in self._registered_LCOWs.values():
-            calculate_variable_from_constraint(var, con)
+    def _add_aggregate_factors(self):
+        pass
